Refactor UserInfo, Footer, Logo
