+++ static/js/components/Footer.jsx
@@ -1,25 +1,22 @@
 import React from 'react';
-import PropTypes from 'prop-types';
+import { useSelector } from 'react-redux';
 
 import styles from './Footer.css';
 
 
-const Footer = ({ version }) => (
-  <div className={styles.footer}>
-    <div className={styles.footerContent}>
-      SkyPortal v{version}. Please file issues at&nbsp;
-      <a href="https://github.com/skyportal/skyportal">
-        https://github.com/skyportal/skyportal
-      </a>
-      .
+const Footer = () => {
+  const version = useSelector(state => state.sysinfo.version);
+  return (
+    <div className={styles.footer}>
+      <div className={styles.footerContent}>
+        SkyPortal v{version}. Please file issues at&nbsp;
+        <a href="https://github.com/skyportal/skyportal">
+          https://github.com/skyportal/skyportal
+        </a>
+        .
+      </div>
     </div>
-  </div>
-);
-Footer.propTypes = {
-  version: PropTypes.string
-};
-Footer.defaultProps = {
-  version: '0.0.0'
+  );
 };
 
 export default Footer;
